Refactor: remove api
